# perf(sequencer): add benchmark for prepare_proposal (ENG-660)