--- a/crates/astria-sequencer/Cargo.toml
+++ b/crates/astria-sequencer/Cargo.toml
@@ -12,7 +12,7 @@ homepage = "https://astria.org"
 name = "astria-sequencer"
 
 [features]
-default = []
+benchmark = ["divan"]
 
 [dependencies]
 astria-core = { path = "../astria-core", features = ["server", "serde"] }
@@ -39,7 +39,7 @@ tower-http = { version = "0.4", features = ["cors"] }
 
 async-trait = { workspace = true }
 bytes = { workspace = true }
-divan = { workspace = true }
+divan = { workspace = true, optional = true }
 futures = { workspace = true }
 hex = { workspace = true, features = ["serde"] }
 ibc-types = { workspace = true, features = ["with_serde"] }
@@ -79,3 +79,4 @@ astria-build-info = { path = "../astria-build-info", features = ["build"] }
 [[bench]]
 name = "benchmark"
 harness = false
+required-features = ["benchmark"]