--- a/crates/astria-sequencer/src/mempool/mod.rs
+++ b/crates/astria-sequencer/src/mempool/mod.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "benchmark")]
 mod benchmarks;
 
 use std::{